Extract resource endpoints to API
