+++ lib/bayberry_web/controllers/main_controller.ex
@@ -1,10 +1,5 @@
 defmodule BayberryWeb.MainController do
   use BayberryWeb, :controller
-  import Application, only: [get_env: 2]
-  alias Bayberry.WordCloud
-
-  @classifier get_env(:bayberry, Bayberry.Service)[:nlp]
-  @redis get_env(:bayberry, Bayberry.Service)[:redis]
 
   def index(conn, _params) do
     render(conn, "index.html")
@@ -22,26 +17,8 @@ defmodule BayberryWeb.MainController do
     render(conn, "nlp.html")
   end
 
-  def sketch(conn, _params) do
-    @redis.json("architecture")
-    |> (&json(conn, &1)).()
-  end
-
-  def topics(conn, _params) do
-    json(conn, @classifier.topics())
-  end
-
   def twitter(conn, _params) do
     render(conn, "twitter.html")
-  end
-
-  def word_cloud(conn, _params) do
-    json(conn, WordCloud.word_count())
-  end
-
-  def world_map(conn, _params) do
-    @redis.json("countries.geo")
-    |> (&json(conn, &1)).()
   end
 
   def blank(conn, _params) do
